+++ config/settings/test.py
@@ -7,17 +7,17 @@ DEBUG = True
 
 # SESSION
 # ------------------------------------------------------------------------------
-SESSIONS["default"]["HOST"] = "TEST_HOST"
-SESSIONS["default"]["PORT"] = "TEST_PORT"
+CONNECTIONS["default"]["HOST"] = "TEST_HOST"
+CONNECTIONS["default"]["PORT"] = "TEST_PORT"
 
-SESSIONS["default"]["SENDER"] = "SENDER_ID"
-SESSIONS["default"]["TARGET"] = "TARGET_ID"
+CONNECTIONS["default"]["SENDER"] = "SENDER_ID"
+CONNECTIONS["default"]["TARGET"] = "TARGET_ID"
 
-SESSIONS["default"]["USERNAME"] = "TEST_USER"
-SESSIONS["default"]["PASSWORD"] = "TEST_PASSWORD"
+CONNECTIONS["default"]["USERNAME"] = "TEST_USER"
+CONNECTIONS["default"]["PASSWORD"] = "TEST_PASSWORD"
 
 # REPEATING GROUPS
-SESSIONS["default"]["GROUP_TEMPLATES"] = {
+CONNECTIONS["default"]["GROUP_TEMPLATES"] = {
     # Routing IDs
     Tag.NoRoutingIDs: [Tag.RoutingType, Tag.RoutingID]
 }